fix: Refactored stat-service - Renamed packages, - Updated pom files for patent its submodules, - Added persistence annotations to RequestInfo
