+++ stats-service/stats-dto/src/main/java/ru/practicum/dto/RequestInfoSummaryDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class RequestInfoSummaryDto {
     private String app;
     private String uri;
-    private Integer hits;
+    private Long hits;
 }
