+++ stats-service/stats-server/src/main/java/ru/practicum/server/service/StatisticsService.java
@@ -1,4 +1,4 @@
-package ru.practicum.server.services;
+package ru.practicum.server.service;
 
 import ru.practicum.dto.RequestInfoDto;
 import ru.practicum.dto.RequestInfoSummaryDto;
